(PC-27889) feat(OfferSummaryInfoList): move the logic from the component inside a hook
