+++ src/features/offerv2/components/OfferContent/OfferContent.tsx
@@ -10,6 +10,7 @@ import { OfferWebMetaHeader } from 'features/offer/components/OfferWebMetaHeader
 import { getOfferPrices } from 'features/offer/helpers/getOfferPrice/getOfferPrice'
 import { useOfferBatchTracking } from 'features/offer/helpers/useOfferBatchTracking/useOfferBatchTracking'
 import { useOfferPlaylist } from 'features/offer/helpers/useOfferPlaylist/useOfferPlaylist'
+import { useOfferSummaryInfoList } from 'features/offer/helpers/useOfferSummaryInfoList/useOfferSummaryInfoList'
 import { OfferAbout } from 'features/offerv2/components/OfferAbout/OfferAbout'
 import { OfferArtists } from 'features/offerv2/components/OfferArtists/OfferArtists'
 import { OfferCTAButton } from 'features/offerv2/components/OfferCTAButton/OfferCTAButton'
@@ -48,6 +49,8 @@ export const OfferContent: FunctionComponent<Props> = ({ offer, searchGroupList,
   const tags = getOfferTags(subcategory.appLabel, extraData)
   const artists = getOfferArtists(subcategory.categoryId, offer)
   const prices = getOfferPrices(offer.stocks)
+
+  const { summaryInfoItems } = useOfferSummaryInfoList({ offer })
 
   const {
     sameArtistPlaylist,
@@ -123,8 +126,13 @@ export const OfferContent: FunctionComponent<Props> = ({ offer, searchGroupList,
           <OfferPrice prices={prices} />
 
           <GroupWithoutGap>
-            {offer.venue.isPermanent ? <OfferVenueButton venue={offer.venue} /> : null}
-            <OfferSummaryInfoList offer={offer} />
+              {!offer.venue.isPermanent && summaryInfoItems.length === 0 ? null : (
+                <Separator.Horizontal testID="topSeparator" />
+              )}
+
+              {summaryInfoItems.length === 0 ? null : (
+                <OfferSummaryInfoList summaryInfoItems={summaryInfoItems} />
+              )}
           </GroupWithoutGap>
 
           <OfferAbout offer={offer} />
